+++ examples/lm1b/configs/default.py
@@ -37,7 +37,7 @@ def get_config():
   config.eval_split = "test"
 
   # Per device batch size for training.
-  config.per_device_batch_size = 128
+  config.per_device_batch_size = 32
 
   # Per device batch size for training.
   config.eval_per_device_batch_size = 32
@@ -58,7 +58,7 @@ def get_config():
   config.num_predict_steps = -1
 
   # Base learning rate.
-  config.learning_rate = 0.05
+  config.learning_rate = 0.0016
 
   # Linear learning rate warmup.
   config.warmup_steps = 1000