+++ tests/linen/module_test.py
@@ -18,8 +18,6 @@ import dataclasses
 import functools
 import operator
 
-
-
 from absl.testing import absltest
 
 import jax